+++ frame/api/app.py
@@ -1,23 +1,20 @@
-from typing import List
-
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 from fastapi_utils.tasks import repeat_every
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
-from fastapi import Depends, FastAPI, HTTPException
 
 from frame import __version__
 from frame.utils import get_logger
 from frame.models.base import SessionLocal
-from frame.exceptions import StationDoesNotExist
-from frame.api.schemas import stations as station_schemas
 from frame.api.services import stations as station_service
+from frame.api.namespaces.stations import router as stations_router
 
 logger = get_logger(__name__)
 
 app = FastAPI(
     title="Frame - BicisBA API",
     description="Stations, status and predictions for the EcoBici system in Buenos Aires.",
+    version=__version__,
 )
 app.add_middleware(
     CORSMiddleware,
@@ -27,15 +24,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 app.add_middleware(GZipMiddleware, minimum_size=256)
-
-
-# Dependency
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
+app.include_router(stations_router)
 
 
 @app.on_event("startup")
@@ -54,43 +43,3 @@ def refresh_stations_status() -> None:
     db = SessionLocal()
     station_service.update_stations_status(db)
     db.close()
-
-
-@app.get("/stations", response_model=List[station_schemas.Station])
-def get_stations(db: Session = Depends(get_db)):
-    return station_service.get_stations(db)
-
-
-@app.get("/stations/status", response_model=List[station_schemas.StationStatus])
-def get_stations_status(db: Session = Depends(get_db)):
-    return station_service.get_stations_status(db)
-
-
-@app.get("/stations/{station_id}", response_model=station_schemas.Station)
-def get_station(station_id: int, db: Session = Depends(get_db)):
-    try:
-        return station_service.get_station(station_id, db)
-    except StationDoesNotExist:
-        raise HTTPException(status_code=404, detail="Station does not exist")
-
-
-@app.get("/stations/{station_id}/status", response_model=station_schemas.StationStatus)
-def get_station_status(station_id: int, db: Session = Depends(get_db)):
-    try:
-        return station_service.get_station_status(station_id, db)
-    except StationDoesNotExist:
-        raise HTTPException(status_code=404, detail="Station does not exist")
-
-
-@app.post(
-    "/stations/{station_id}/prediction", response_model=station_schemas.Prediction
-)
-def predict_for_station(
-    station_id: int,
-    prediction_params: station_schemas.PredictionParams,
-    db: Session = Depends(get_db),
-):
-    try:
-        return station_service.predict(station_id, prediction_params, db)
-    except StationDoesNotExist:
-        raise HTTPException(status_code=404, detail="Station does not exist")