+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.5] - 2022-10-11
+### Fixed
+- Logging through Uvicorn
+- Station ids as integers from the API fetch
+
+### Changed
+- Updating stations info and status with the stations service
+- Namespaces for OpenAPI
+
+### Added
+- CLI command to run the server
+
 ## [0.4.4] - 2022-10-11
 ### Fixed
 - Avoid integrity errors on stations status update